// ase-webclient/src/app/model/userdata.ts
/**
 * Created by Matthias on 30.11.17.
 */
export interface UserData {
  /**
   * The first name of the student
   */
  firstName: string;

  /**
   * The last name of the student
   */
  lastName: string;

  /**
   * The user's email account
   */
  email: string;

  /**
   * The Id of the user
   */
  id: number;

  password: string;

}